Refactored lib structure
