+++ learnrl/evaluation.py
@@ -5,8 +5,8 @@
 Evaluation methodes to modify the value fonctions from experiences
 """
 from learnrl.core import Memory
-from learnrl.agent_parts.control import Control
-from learnrl.agent_parts.estimator import Estimator
+from learnrl.control import Control
+from learnrl.estimator import Estimator
 
 import numpy as np
 from copy import copy
